refactor: move votingNeuronSelectStore
