+++ frontend/src/tests/lib/stores/proposals.store.spec.ts
@@ -4,8 +4,6 @@ import {
   proposalsFiltersStore,
   proposalsStore,
 } from "$lib/stores/proposals.store";
-import { votingNeuronSelectStore } from "$lib/stores/vote-registration.store";
-import type { VotingNeuron } from "$lib/types/proposals";
 import { generateMockProposals } from "$tests/mocks/proposal.mock";
 import { ProposalRewardStatus, ProposalStatus, Topic } from "@dfinity/nns";
 import { get } from "svelte/store";
@@ -186,55 +184,6 @@ describe("proposals-store", () => {
         topics: filter,
       });
       expect(filters.lastAppliedFilter).toBeUndefined();
-    });
-  });
-
-  describe("votingNeuronSelectStore", () => {
-    const neuronIds = [0, 1, 2].map(String);
-    const neurons = neuronIds.map(
-      (neuronIdString) =>
-        ({
-          neuronIdString,
-          votingPower: 0n,
-        } as VotingNeuron)
-    );
-
-    it("should set neurons", () => {
-      votingNeuronSelectStore.set(neurons);
-      expect(get(votingNeuronSelectStore).neurons).toEqual(neurons);
-    });
-
-    it("should select all on set", () => {
-      votingNeuronSelectStore.set(neurons);
-      expect(get(votingNeuronSelectStore).selectedIds).toEqual(
-        neuronIds.map(String)
-      );
-    });
-
-    it("should preserve user selection", () => {
-      votingNeuronSelectStore.set(neurons);
-      votingNeuronSelectStore.toggleSelection(`${neuronIds[1]}`);
-      votingNeuronSelectStore.updateNeurons([
-        ...neurons,
-        {
-          neuronIdString: "3",
-          votingPower: 0n,
-        } as VotingNeuron,
-      ]);
-      expect(get(votingNeuronSelectStore).selectedIds).toEqual(
-        [0, 2, 3].map(String)
-      );
-    });
-
-    it("should toggle by neuronId", () => {
-      votingNeuronSelectStore.set(neurons);
-
-      votingNeuronSelectStore.toggleSelection(`${neuronIds[1]}`);
-      votingNeuronSelectStore.toggleSelection(`${neuronIds[1]}`);
-      votingNeuronSelectStore.toggleSelection(`${neuronIds[2]}`);
-      expect(get(votingNeuronSelectStore).selectedIds).toEqual(
-        [neuronIds[0], neuronIds[1]].map(String)
-      );
     });
   });
